# Senac-SenacBikeStore/app/Http/Requests/v1/StoreCategoriaRequest.php
<?php

namespace App\Http\Requests\v1;

use Illuminate\Foundation\Http\FormRequest;

class StoreCategoriaRequest extends FormRequest
{
    /**
     * Determine if the user is authorized to make this request.
     */
    public function authorize(): bool
    {
        return true;
    }

    /**
     * Get the validation rules that apply to the request.
     *
     * @return array<string, \Illuminate\Contracts\Validation\Rule|array|string>
     */
    public function rules(): array
    {
        return [
            "nome_da_categoria" => "required"
        ];
    }
}//quando recebr requisição deve-se passar um json com esses dados